backend/requirements/deploy.txt
# ============================================
# Deployment Requirements (Render Free Tier)
# ============================================
# 최적화 목표:
# - RAM < 512MB (Render Free Tier 제한)
# - CPU < 0.1 vCPU
# - 빌드 시간 < 1분
# - 완전 무료 배포

# ============================================
# FastAPI and ASGI
# ============================================
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anyio>=4.8.0,<5.0.0

# ============================================
# Database
# ============================================
sqlalchemy>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0  # PostgreSQL (Supabase)

# ============================================
# Redis (Optional for production)
# ============================================
redis>=5.0.1

# ============================================
# Authentication
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6

# ============================================
# HTTP Client
# ============================================
aiohttp>=3.10.0
urllib3>=1.24.2,<2.4.0

# ============================================
# Data Validation
# ============================================
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ============================================
# Environment Variables
# ============================================
python-dotenv>=1.0.0

# ============================================
# Logging
# ============================================
loguru>=0.7.2

# ============================================
# Background Tasks
# ============================================
celery>=5.3.4
apscheduler>=3.10.4

# ============================================
# Web Scraping
# ============================================
beautifulsoup4>=4.12.2
selenium>=4.15.2
lxml>=4.9.3

# ============================================
# Date/Time
# ============================================
python-dateutil>=2.8.2

# ============================================
# Utilities
# ============================================
python-slugify>=8.0.1

# ============================================
# AI - Render Free Tier Optimized
# ============================================

# Gemini API (LLM + Embedding)
google-genai>=1.0.0

# Lightweight ML
numpy>=1.26.0        # 코사인 유사도 계산
pandas>=2.1.0        # 데이터 처리
tqdm>=4.66.0         # 진행 표시

# ============================================
# Monitoring (Optional)
# ============================================
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0


